Deduplicate rabbitmq controller code using `ResourceBuilder` (#300)
This PR aims to make the implementation of rabbitmq controller better
arranged.

The first thing we do is to make the reconcile step type
(`RabbitmqReconcileStep`) more expressive: instead of literally
concatenating the request type and resource type, we split them and use
two fields for the step. The program can't tell what the previous step
type represents so we have to exhaustively handle all cases which
introduce much duplications. But now we can talk about what the request
type is and what the resource object type created/updated/got is.

Then we can independently handle the creation/update process of each
kind of object in their builder, which means each type of resource
object will have its owner `ResourceBuilder`. Ideally, `ResourceBuilder`
should be able to:

- Receive a cr and an old object (of `DynamicObject` type) and return an
updated object or error;
- Receive a cr and return a brand new object or error.
- Return the basic information of the object: name, namespace,
api_version, etc.
Those methods all handle dynamic objects instead of the unmarshaled
resource object so that: (1) we don't have need an extra get result
checker (and actually the get result checker is for update) and can
simply do it checking together with the check before update; (2) it
reduced the layers of checking results when processing k8s get response;
(3) we don't need to marshal the updated/created object in the
reconciler which is not convenient.

What the next step is should be bound to the step and decided by the
reconciler; put it in another word, the resource builder should only
care about the very object and should not be aware of what the next
object to process or the subsequent reconcile state. But currently we
have a method `state_after_create_or_update` which is against such a
principle. We may refine it later.

---------

Signed-off-by: Wenjie Ma <rebekah1368@gmail.com>