Fix oscillation problem by renaming the client service of rabbitmq (#335)
This is to avoid a naming conflict bug when there are multiple rabbitmq
cluster (cr) instances. Suppose one instance is called `foo` and the
other is called `foo-nodes`. Before this change, the headless service of
the first cr will be called `foo-nodes` and the client service of the
second cr will be called `foo-nodes` as well. Thus, the reconciler
instances for the two crs will start to compete on the `foo-nodes`
service, causing oscillation problems and liveness violations.

This renaming fixes the problem. From now on, for every rabbitmq cluster
instance, its headless and client services will never collide with the
services of other rabbitmq cluster instances.

---------

Signed-off-by: Wenjie Ma <rebekah1368@gmail.com>