Specify stateful set and daemon set controller (#322)
This PR specifies the behavior of two important built-in controllers:
stateful set controller and daemon set controller, which continuously
updates the status of any stateful or daemon set object. Here we don't
assume the stateful or daemon set pods eventually all become ready;
instead, our specification says the stateful or daemon set controller
updates the `.status.ready_replicas` (for stateful set) or
`.status.number_ready` (for daemon set) to some value.

A problem introduced by the two new built-in controllers is that, if
they continuously update the object's status (which increments the
resource version number), the custom controller will never be able to
update the same object. That is, there exists some execution where the
custom controller keeps losing the race on a stateful or a daemon set
object to the built-in controller, where liveness is impossible.

To solve the problem, we add one more constraint (assumption) to the two
built-in controllers: they eventually stop editing the status of an
object until that object's spec gets updated again. This assumption is
encoded in the stabilizer state machine (as part of the built-in
controllers state machine), which picks an object and adds the object
into a stable set. The two built-in controllers only update the status
of an object when it is not present in the stable set. Besides, every
time the api server creates or updates an object, that object will be
removed from the stable set (no-op update won't remove any object from
the stable set).

With all the specifications above, we derive a lemma that helps liveness
proof even in the presence of built-in controllers
```
pub proof fn lemma_true_leads_to_always_object_not_exist_or_updated_or_no_more_pending_req(
    spec: TempPred<Self>, key: ObjectRef, requirements: MutableByUpdatePred
)
    requires
        spec.entails(always(lift_action(Self::next()))),
        spec.entails(tla_forall(|i| Self::kubernetes_api_next().weak_fairness(i))),
        spec.entails(tla_forall(|i| Self::builtin_controllers_next().weak_fairness(i))),
        spec.entails(always(lift_state(Self::every_in_flight_create_req_msg_for_this_object_satisfies(key, requirements)))),
        spec.entails(always(lift_state(Self::every_in_flight_update_req_msg_for_this_object_satisfies(key, requirements)))),
    ensures
        spec.entails(true_pred().leads_to(always(lift_state(
            |s: Self| {
                ||| !s.resources().contains_key(key)
                ||| s.resources().contains_key(key) && requirements(s.resources()[key].mutable_subset())
                ||| {
                    &&& Self::no_status_update_req_msg_from_bc_for_this_object(key)(s)
                    &&& s.stable_resources().contains(key)
                }
            }
        )))),
```
Informally, the lemma says that assuming all the create or update
requests to this object (of key) satisfy the same `requirements`,
eventually, it is always true that: (1) the object does not exist, or
(2) the object exists and satisfies the `requirements`, and (3) there is
no more update-status request for this object.

This lemma is supposed to be used as one of the phased invariants in our
liveness proof. It will be applied when proving the update request sent
by the custom controller can update the object to the desired state: if
(2) is true, then the object is already in the desired state and we
don't need to care about whether this request goes through or not, and
if (3) is true then the update request can easily go through without
racing with update-status requests from built-in controllers (here (1)
is trivially false).

A limitation of this lemma is that applying it only shows us the object
eventually gets to the desired state, but it does not show the update
request returns with OK. If any operation happens only after the update
stateful set request returns with OK, then we will need another lemma
saying that the update-status requests eventually are all gone as long
as the object never gets updated (so it is never removed from the stable
set). We will work on such a lemma when we face this problem.

---------

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>