Add a prototype of verified ReplicaSet controller (#480)
The prototype replica set controller implements the basic scaling
up/down functionality. It lists all the pods in the same namespace,
filters out the pods that do not match the replica set object's
selector, and them create/delete pods to make sure there are `replicas`
number of matched pods.

The replica set controller is not verified yet, and there are still more
features to implement, such as claiming orphan pods.

---------

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>