Add liveness specification for VReplicaSet controller (#497)
Here I state a basic liveness theorem, based on Eventually Stable
Reconciliation, for ReplicaSets. To check whether the cluster state
matches the spec, I gather up all pods that match the label selector
`selector` and are owned by the ReplicaSet into a Verus set. I then
check whether the cardinality of that set is equal to the `replicas`
field.

---------

Signed-off-by: Cody Rivera <codyjr3@illinois.edu>