Support `generate_name` in the API server model (#499)
This PR revises the API server model to allow resource creation requests
that carry a `generate_name` rather than a `name`. For such requests,
the API server will generate a unique name for the object and create the
object with the generated name. Before the revision, such requests are
rejected by our model. We do not specify the details on how the API
server guarantees the uniqueness of the generated name, but rather use a
trusted proof function saying that the generated name is different from
any existing names. Note that this trusted proof is better written as a
"spec ensures" once that gets supported by Verus.

We make this revision mainly to enable verifying Kubernetes core
controllers such as the ReplicaSet controller that often creates pods
with `generate_name`.

This revision breaks many proofs of the custom controllers we have
built, but it is straightforward to fix these broken proofs. In most
cases, we only need to say that there does not exist any create request
that might create the "same" object.

---------

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>