Add a toy example for controller interaction (#501)
This PR adds two toy controllers: producer and consumer to illustrate
how controllers work together. Their collaboration is simple: the
consumer waits for the producer to create a pod and then the consumer
adds a label to the pod.

The implementation is done but the model and proof are not.

---------

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>