Simplify compositional theorems by using `Map` to represent controllers (#512)
This PR uses `Map` instead of `Seq` to store the controllers in the
cluster. The major benefit is that `Map` makes it much easier to state
"controllers other than x, y, and z satisfy properties P" by simply
removing the entries of x, y, and z from the map and applying P to the
resulting map.

One limitation of using `Map` is that we cannot have multiple instances
of the same controllers in the cluster. However, in practice this is not
a concern: there is no reason to have two identical controllers running
in the cluster. The two will conflict and interfere with each other. The
common practice is to have only one active instance (the primary) of
each controller running, with multiple threads internally handling
different pieces of the cluster state, and some other inactive instances
as backup if the primary one terminates. This behavior is already well
modeled in our state machine.

---------

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>