Compose controllers incrementally (#515)
This PR introduces a way for developers to compose controllers
incrementally when proving the correctness of a cluster of controllers.
By "incrementally", we mean that if we already have a proof `P` for the
correctness of a controller `foo` and its producers, we can develop
another proof for the correctness of controller `bar` that relies on
`foo` and its producers on top of `P` by focusing on reasoning about
`bar`.

More concretely, this PR introduces:
- `BaseComposition`: a trait with five proof functions (as trait
methods) as the proof obligations to compose a consumer and a sequence
of producers.
- `IncrementalComposition`: a trait with three proof functions as the
proof obligations to incrementally compose a consumer with a sequence of
producers (that have already been proved correct).
- Two examples of how to use `BaseComposition` and
`IncrementalComposition`.

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>