Differentiate between different custom resource types (#520)
This PR modifies the `Kind` type to let its `CustomResourceKind` variant
carry a `StringView`, which can be used to differentiate between
different custom resource types. For example, the ZooKeeper controller's
CR type should be `CustomResourceKind("zookeeper"@)` while the RabbitMQ
controller's should be `CustomResourceKind("rabbitmq"@)`. This is
important for state machine v2 because it will take in multiple
controllers and multiple custom resource types and we need to
differentiate them during the proof.

---------

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>