Add PodEvent state machine to model random pod API requests (#522)
I add a state machine to the cluster called `PodEvent` which issues
random pod create, delete, and update requests on pods, while it is
enabled. I also model a transition which disables `PodEvent`.

However, as implemented, not only are the liveness proofs of the
controllers broken (that is to be expected), but also some of the proofs
involving the builtin controllers in the compound state machine. For
example,
https://github.com/anvil-verifier/anvil/blob/3d0e043c1227c31213d9d07c050e1de32db8b387/src/kubernetes_cluster/proof/daemon_set_controller.rs#L298C10-L298C89
does not go through. Similar proofs are marked with a comment.

---------

Signed-off-by: Cody Rivera <codyjr3@illinois.edu>