Support `preconditions` for Delete request (#545)
This PR introduces `preconditions` of Kubernetes delete requests. The
`preconditions` contain a uid and a resource version. When handling the
delete request, the API server checks whether the object's uid and
resource version match the ones provided by the request and returns
`Conflict` error if not. Kubernetes implements this feature here:
https://github.com/kubernetes/kubernetes/blob/v1.30.0/staging/src/k8s.io/apiserver/pkg/storage/interfaces.go#L140-L153.
The `preconditions` is optional. Introducing `preconditions` mainly
requires changes to the API server model (modeling the uid and resource
version checking) and the shim layer (passing the `preconditions` to
`kube`).

This PR also fixes some of the lemmas that are broken due to
`preconditions`.

Having `preconditions` allows us to use more restrictive deletion. This
is useful to ensure no interference between controllers that manage pods
--- when a controller decides to delete a pod that the controller
believes belongs to itself, the controller should pass to the
`preconditions` the resource version of the pod it observes. When API
server handles the delete request, if the pod happens to be owned by a
different controller, then the deletion will be invalidated (because the
resource version must have already changed since the controller's
observation) so that the controller won't interfere with the new owner
of the pod.

---------

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>