Model errors of create with `generate_name` (#566)
By removing the precondition on the API error type from `drop_req`, we
generalize our fault model by including a new type of error: creation
request with `generate_name` fails to create the object when all the
creation attempts conflict with existing objects. This type of error is
non-deterministic because the API server uses a name randomly generated
when handling a creation request whose `name` is empty and
`generate_name` is in use. We treat it as a transient error and assume
that it will eventually disappear. Otherwise, there is no way for
controllers that issue creation requests with `generate_name` to finish
reconciliation.

---------

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>