Restate ESR property for VRS; breaks proofs (#628)
I restate the ESR property as filtering `s.resources().values()` as
suggested in the comments of one of the files.

Note that we define the pods matching a replicaset by filtering
`s.resources().values()` using the function `matching_pods`. My
intention is to have this function replace `matching_pod_keys` (what
used to be called `matching_pods`), which gives a set of _keys_ of
matching pods, and `matching_pod_entries`, which provides a _submap_ of
keys to matching pods.

---------

Signed-off-by: Cody Rivera <codyjr3@illinois.edu>