Support "transactional" Get-then-Update and Get-then-Delete (#621)
This PR introduces "transactional" APIs: Get-then-Update and
Get-then-Delete. This PR breaks some proofs of the VRS controller.
Fixing the broken proof is left for future work.

**Spec of Get-then-Update**: The API server checks whether the object
exists and whether the object is owned by some owner reference, and if
so the API server updates the object using the current resource version
and uid. These operations happen in a single step. Get-then-Update will
never fail due to Conflict error because it applies update using the
current resource version and uid.

**Implementation of Get-then-Update**: Kubernetes does not provide such
a transactional API, so Get-then-Update is implemented by retrying get
and update. It first gets the object, checks its owner, then updates the
object using the resource version and uid from the get result. If the
update fails due to Conflict error (e.g., some other controller updates
the object between our get and update), it retries get and update until
it succeeds or other error happens. Note that the termination of the
implementation depends on fairness.

The spec and implementation of Get-then-Delete are similar to
Get-then-Update.

**Why the implementation refines the spec**: There is currently no
machine-checked proof for the refinement. The proof will require the
fairness condition that guarantees the implementation's termination. The
intuition of the proof is to show that for any possible trace of the
implementation, if the implementation (re)tries "Get-then-Update" for N
times, the first N-1 tries will not cause any changes to the cluster
state (as they failed due to Conflict error) so the first N-1 tries are
mapped to no-ops to the trace of the spec. For the Nth try of
"Get-then-Update", there are two cases to consider: (1) the object gets
deleted between the get and update, and the Nth try fails with
ObjectNotFound error. In this case, we map the implementation's get to a
no-op before the deletion, and the implementation's update to the spec's
atomic get-then-update after the deletion. (2) the object still exists
when the update happens, then it implies no other controllers touches
the object between the get and update (otherwise the Nth try will also
fail with Conflict error). In this case, we map the implementation's
"Get-then-Update" to the spec's atomic Get-then-Update.

**Limitation**: Currently, Get-then-Update and Get-then-Delete's
expressiveness is limited: they only check the object's owner before
issuing update or delete, instead of checking arbitrary conditions of
the object. To support arbitrary condition check, the controller's
`reconcile_core` needs to return a function (e.g., `Fn(DynamicObject) ->
bool`) as a field of the request struct. Rust requires the function to
be wrapped in a `Box<dyn ...>`, but `dyn` is not supported by Verus for
now.

---------

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>