Fix off-by-one error in controller action precondition (#643)
This fixes an off-by-one error in `controller_action_pre`: instead of
forcing the `ObjectRef` to be non-null, it forces the message to be
`null`, whereas it is allowed that a controller takes a step without a
message as input.

This is separate from the rest of subtask in
(https://github.com/xlab-uiuc/anvil/issues/26) since spec/trusted code
is modified.

Signed-off-by: Cody Rivera <codyjr3@illinois.edu>