Fix WF1 proofs on 'top' (liveness) side of Bridge Invariant (#647)
This PR rewrites the liveness proof to use a bridge invariant saying
that no message potentially interfering with the current reconcile for
`vrs` can be in the network, unless the source is precisely the
reconcile for `vrs`. The fixes are complete minus a helper lemma, and
lemmas specifying the results of the API server taking an action.

---------

Signed-off-by: Cody Rivera <codyjr3@illinois.edu>
Signed-off-by: Cody Rivera <cjrivera@wisc.edu>