Change rely-guarantee properties for transactional API (#651)
Here I update the rely and guarantee conditions for `VReplicaSet`
dealing with messages using the transactional API to not explicitly
discuss the state of etcd. Instead, they merely put conditions on the
kind of the owner reference sent along with the request. The API server
will ensure that this owner reference and version number of the object
in etcd matches, aborting otherwise.

These changes are analogous to those already in `VDeployment`. They
break a couple of proofs I was already expecting to delete soon anyway.

---------

Signed-off-by: Cody Rivera <cjrivera@wisc.edu>