Replace `get_Some_0` with `unwrap` in macro (#668)
in #664 I replaced most of `get_Some_0` with `->0`, while those in
macros defined in `src/reconciler/spec/io.rs` caused some error. As
`unwrap` is not deprecated this PR replace them with `unwrap` to reduce
the lengthy deprecation warning in every verification.

However, `is_Some` in macro cannot be replaced in this way and will
still cause warning. But it's less overwhelming with the `unwrap`
replacement.