Introduce valid_owned_object for vd reconciler (#667)
Before this PR `well_formed` is required for VRS objects in VDeployment
controller local state. It's too strong and caused a problem that
`new_vrs` created by `make_replica_set` does not have `metadata.uid |
resource_version` required in `well_formed`.

This PR change the `well_formed` to be `valid_owned_object`.

Minor changes:
- `valid_owned_object` additionally carries requirements on
`vrs.metadata.owner_references` which is guaranteed to be satisfied
throughout the reconciliation to simplify the reasoning.
- Use `valid_owned_object` to replace `filter_vrs_list` which is
inconvenient to reason compared with the filter