Add temporal rule `strengthen_leads_to_with_until` (#698)
`strengthen_leads_to_with_until` is useful for reasoning about the
pattern where controller A creates object x, waiting for object y to be
created, and controller B implements the spec that if x always exists,
then it eventually creates y.

Note that this lemma allows controller A to update or delete x after y
is created by B, which happens in real-world controllers like
Kubernetes' Deployment. To meet the requirements of maxSurge and
maxUnavailable, Deployment controller, after creating a ReplicaSet
object, might wait for the ReplicaSet controller to create pods, and
then it will update the ReplicaSet object again.

This lemma allows compositional proof in such a case. When verifying the
ReplicaSet controller, one can assume `[]p` and focus on proving
eventually `q` holds. When composing it with the Deployment controller,
although `[]p` is no longer true, we can still prove eventually `q`
holds, as long as we can show that the Deployment controller (and other
controllers) preserves `p` until `q` holds. We don't have until operator
in TLA, so we formalize the until argument using `[](p /\ next => p' \/
q')`.

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>