Model VStatefulSet controller's version update logic (#701)
Version update logic is implemented by simply deleting the unmatched pod
with the largest ordinal `n`, and then finishing the reconcile(). The
remaining job will be done by the next reconcile() that will create a
new pod with ordinal `n` with the new template, and then delete the next
outdated pod.

---------

Signed-off-by: Xudong Sun <xudongs3@illinois.edu>