Implement vstatefulset_controller (#706)
This PR implements the exec code for a verified StatefulSet controller
(in `src/controllers/vstateful_controllers/exec/reconciler.rs`),
implementing the spec in
(`src/controllers/vstateful_controllers/model/reconciler.rs`). While the
entire spec has been implemented completely, a couple of things are left
as TODOs for future PRs:

- The conformance proofs for `update_storage`, `handle_delete_outdated`,
and `get_largest_ordinal_of_unmatched_pods` need to be completed.
- There is a trusted `sort_pods_by_ord` function in
src/controllers/vstatefulset_controller/trusted/reconciler.rs. I did not
implement a verified version of this for now since it's quite difficult
to implement the verified sort needed. Instead, I just trust the Rust
standard library's `Vec::sort_by`. We may want to replace this with an
actual verified sort in the future.
- The implementation currently fails the admission test. I need to
investigate why this is failing and fix this. Since the fix is likely
out of scope for this PR, I have masked out the vstatefulset admission
test.